feat: Add getListStudents function for retrieving a list of students
This commit adds a new file, `0-get_list_students.js`, which contains the implementation of the `getListStudents` function. This function returns an array of student objects with their respective IDs, first names, and locations.

The purpose of this change is to provide a convenient way to retrieve a list of students for further data manipulation in JavaScript ES6.