feat: Add getListStudentIds function for retrieving a list of student IDs
This commit adds a new file, `1-get_list_student_ids.js`, which contains the implementation of the `getListStudentIds` function. This function takes an array of student objects and returns an array of their respective IDs.

The purpose of this change is to provide a convenient way to retrieve a list of student IDs for further data manipulation in JavaScript ES6.