feat: Implement getStudentIdsSum function for calculating the sum of student IDs
This commit adds a new file, `3-get_ids_sum.js`, which contains the implementation of the `getStudentIdsSum` function. This function takes an array of student objects and returns the sum of their respective IDs.

The purpose of this change is to provide a convenient way to calculate the sum of student IDs for data analysis or other purposes in JavaScript ES6.