feat: Update student grades by city
This commit adds the `updateStudentGradeByCity` function to the `ES6_data_manipulation/4-update_grade_by_city.js` file. This function takes an array of student objects, a city, and an array of new grades. It filters the students based on the provided city and updates their grades according to the new grades array.

The purpose of this change is to provide a convenient way to update student grades based on their location in JavaScript ES6.