feat: Add groceriesList function for creating a Map of groceries
This commit adds the `groceriesList` function to the `ES6_data_manipulation/9-groceries_list.js` file. This function returns a new Map object that contains a list of groceries and their quantities.

The purpose of this change is to provide a convenient way to create and manage a grocery list using a Map in JavaScript ES6.