feat: Add signUpUser function for signing up users
This commit adds the `signUpUser` function to the `ES6_promise/4-user-promise.js` file. The function takes in a user's first name and last name as parameters and returns a Promise that resolves to an object containing the user's first name and last name.

The purpose of this change is to provide a function for signing up users by creating a Promise that resolves with the user's information.