feat: Add divideFunction for dividing two numbers
The commit adds the `divideFunction` to the `ES6_promise/8-try.js` file. This function takes in a numerator and denominator as parameters and returns the result of dividing the numerator by the denominator. If the denominator is 0, an error is thrown with the message "cannot divide by 0".

The purpose of this change is to provide a function for dividing two numbers and handling the case of dividing by 0.