feat: Add guardrail function for handling errors in mathFunction
The commit adds the `guardrail` function to the `ES6_promise/9-try.js` file. This function takes in a `mathFunction` as a parameter and executes it within a try-catch block. If the `mathFunction` throws an error, the error message is pushed to the `queue` array. Otherwise, the result of the `mathFunction` is pushed to the `queue` array. Finally, the string "Guardrail was processed" is pushed to the `queue` array and the `queue` is returned.

The purpose of this change is to provide a function for handling errors that may occur when executing the `mathFunction`.